Broke apart presence code into multiple files as per style guide
