+++ src/IO.Ably.Shared/IO.Ably.Shared.projitems
@@ -16,6 +16,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)ApiKey.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PaginatedRequestParams.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ChannelEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Realtime\Presence\PresenceManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Realtime\Presence\PresenceMap.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Realtime\Presence\QueuedPresenceMessage.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Types\AuthDetails.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)AuthMethod.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)AuthOptions.cs" />
@@ -83,7 +86,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\Handlers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\MessageHandlerAction.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\IRealtimeChannel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Realtime\Presence.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Realtime\Presence\GetOptions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\RealtimeTransportData.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\Channel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\ChannelOptions.cs" />
