+++ src/IO.Ably.Shared/Realtime/Presence/PresenceManager.cs
@@ -1,25 +1,14 @@
 ﻿using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using IO.Ably;
 using IO.Ably.Transport;
 using IO.Ably.Types;
 
-namespace IO.Ably.Realtime
+namespace IO.Ably.Realtime.Presence
 {
-    public class GetOptions
-    {
-        public bool WaitForSync { get; set; } = true;
-
-        public string ClientId { get; set; }
-
-        public string ConnectionId { get; set; }
-    }
-
-    public class Presence : IDisposable
+    public partial class PresenceManager : IDisposable
     {
         internal ILogger Logger { get; private set; }
 
@@ -60,7 +49,7 @@ namespace IO.Ably.Realtime
             SyncComplete = true;
         }
 
-        internal Presence(IConnectionManager connection, RealtimeChannel channel, string cliendId, ILogger logger)
+        internal PresenceManager(IConnectionManager connection, RealtimeChannel channel, string cliendId, ILogger logger)
         {
             Logger = logger;
             Map = new PresenceMap(channel.Name, logger);
@@ -363,10 +352,10 @@ namespace IO.Ably.Realtime
             var residualMembers = Map.EndSync();
 
             /*
-                         * RTP19: ... The PresenceMessage published should contain the original attributes of the presence
-                         * member with the action set to LEAVE, PresenceMessage#id set to null, and the timestamp set
-                         * to the current time ...
-                         */
+             * RTP19: ... The PresenceMessage published should contain the original attributes of the presence
+             * member with the action set to LEAVE, PresenceMessage#id set to null, and the timestamp set
+             * to the current time ...
+             */
             foreach (var presenceMessage in residualMembers)
             {
                 presenceMessage.Action = PresenceAction.Leave;
@@ -533,219 +522,9 @@ namespace IO.Ably.Realtime
             Map.StartSync();
         }
 
-        internal class PresenceMap
-        {
-            internal ILogger Logger { get; private set; }
-
-            internal event EventHandler SyncNoLongerInProgress;
-
-            private readonly string _channelName;
-            private readonly object _lock = new object();
-
-            public enum State
-            {
-                Initialized,
-                SyncStarting,
-                InSync,
-                Failed
-            }
-
-            private readonly ConcurrentDictionary<string, PresenceMessage> _members;
-            private ICollection<string> _residualMembers;
-            private bool _isSyncInProgress;
-
-            public PresenceMap(string channelName, ILogger logger)
-            {
-                Logger = logger;
-                _channelName = channelName;
-                _members = new ConcurrentDictionary<string, PresenceMessage>();
-            }
-
-            public bool IsSyncInProgress
-            {
-                get => _isSyncInProgress;
-                private set
-                {
-                    var previous = _isSyncInProgress;
-                    _isSyncInProgress = value;
-
-                    // if we have gone from true to false then fire SyncNoLongerInProgress
-                    if (previous && !_isSyncInProgress)
-                    {
-                        OnSyncNoLongerInProgress();
-                    }
-                }
-            }
-
-            public bool InitialSyncCompleted { get; private set; }
-
-            public PresenceMessage[] Values
-            {
-                get
-                {
-                    return _members.Values.Where(c => c.Action != PresenceAction.Absent)
-                        .ToArray();
-                }
-            }
-
-            public bool Put(PresenceMessage item)
-            {
-                lock (_lock)
-                {
-                    // we've seen this member, so do not remove it at the end of sync
-                    _residualMembers?.Remove(item.MemberKey);
-                }
-
-                try
-                {
-                    if (_members.TryGetValue(item.MemberKey, out var existingItem) && existingItem.IsNewerThan(item))
-                    {
-                        return false;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Logger.Error($"PresenceMap.Put | Channel: {_channelName}, Error: {ex.Message}");
-                    throw ex;
-                }
-
-                switch (item.Action)
-                {
-                    case PresenceAction.Enter:
-                    case PresenceAction.Update:
-                        item = item.ShallowClone();
-                        item.Action = PresenceAction.Present;
-                        break;
-                }
-
-                _members[item.MemberKey] = item;
-
-                return true;
-            }
-
-            public bool Remove(PresenceMessage item)
-            {
-                PresenceMessage existingItem;
-                if (_members.TryGetValue(item.MemberKey, out existingItem) && existingItem.IsNewerThan(item))
-                {
-                    return false;
-                }
-
-                _members.TryRemove(item.MemberKey, out PresenceMessage _);
-                if (existingItem?.Action == PresenceAction.Absent)
-                {
-                    return false;
-                }
-
-                return true;
-            }
-
-            public void StartSync()
-            {
-                if (Logger.IsDebug)
-                {
-                    Logger.Debug($"StartSync | Channel: {_channelName}, SyncInProgress: {IsSyncInProgress}");
-                }
-
-                if (!IsSyncInProgress)
-                {
-                    lock (_lock)
-                    {
-                        _residualMembers = new HashSet<string>(_members.Keys);
-                        IsSyncInProgress = true;
-                    }
-                }
-            }
-
-            public PresenceMessage[] EndSync()
-            {
-                if (Logger.IsDebug)
-                {
-                    Logger.Debug($"EndSync | Channel: {_channelName}, SyncInProgress: {IsSyncInProgress}");
-                }
-
-                List<PresenceMessage> removed = new List<PresenceMessage>();
-                try
-                {
-                    if (!IsSyncInProgress)
-                    {
-                        return removed.ToArray();
-                    }
-
-                    // We can now strip out the ABSENT members, as we have
-                    // received all of the out-of-order sync messages
-                    foreach (var member in _members.ToArray())
-                    {
-                        if (member.Value.Action == PresenceAction.Absent)
-                        {
-                            _members.TryRemove(member.Key, out PresenceMessage _);
-                        }
-                    }
-
-                    lock (_lock)
-                    {
-                        if (_residualMembers != null)
-                        {
-                            // Any members that were present at the start of the sync,
-                            // and have not been seen in sync, can be removed
-                            foreach (var member in _residualMembers)
-                            {
-                                if (_members.TryRemove(member, out PresenceMessage pm))
-                                {
-                                    removed.Add(pm);
-                                }
-                            }
-
-                            _residualMembers = null;
-                        }
-
-                        IsSyncInProgress = false;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Logger.Error($"PresenceMap.EndSync | Channel: {_channelName}, Error: {ex.Message}");
-                    throw ex;
-                }
-                finally
-                {
-                    InitialSyncCompleted = true;
-                }
-
-                return removed.ToArray();
-            }
-
-            public void Clear()
-            {
-                lock (_lock)
-                {
-                    _members?.Clear();
-                    _residualMembers?.Clear();
-                }
-            }
-
-            protected virtual void OnSyncNoLongerInProgress()
-            {
-                SyncNoLongerInProgress?.Invoke(this, EventArgs.Empty);
-            }
-        }
-
         protected virtual void OnInitialSyncCompleted()
         {
             InitialSyncCompleted?.Invoke(this, EventArgs.Empty);
         }
-    }
-
-    internal class QueuedPresenceMessage
-    {
-        public QueuedPresenceMessage(PresenceMessage message, Action<bool, ErrorInfo> callback)
-        {
-            Message = message;
-            Callback = callback;
-        }
-
-        public PresenceMessage Message { get; }
-
-        public Action<bool, ErrorInfo> Callback { get; }
     }
 }
