+++ src/IO.Ably.Shared/Realtime/Presence/QueuedPresenceMessage.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace IO.Ably.Realtime
+{
+    internal class QueuedPresenceMessage
+    {
+        public QueuedPresenceMessage(PresenceMessage message, Action<bool, ErrorInfo> callback)
+        {
+            Message = message;
+            Callback = callback;
+        }
+
+        public PresenceMessage Message { get; }
+
+        public Action<bool, ErrorInfo> Callback { get; }
+    }
+}
